requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0

# Spotify API
spotipy>=2.22.0

# Web framework
streamlit>=1.25.0
plotly>=5.15.0
altair>=4.2.0

# Data processing
pyarrow>=10.0.0
fastparquet>=0.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0
joblib>=1.2.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional: Advanced ML
optuna>=3.0.0  # For hyperparameter optimization
shap>=0.41.0   # For model interpretability
